+++ Core/Simulation/OffSpecSimulation.h
@@ -39,7 +39,7 @@ public:
     void prepareSimulation() final;
 
     //! Gets the number of elements this simulation needs to calculate
-    int numberOfSimulationElements() const final;
+    size_t numberOfSimulationElements() const final;
 
     //! Returns clone of the detector intensity map
     OutputData<double>* getDetectorIntensity(
